Make tests run in CI
