--- .circleci/devcluster/elastic-base-s3.devcluster.yaml
@@ -1,22 +1,0 @@
-stages:
-  - db:
-      name: db
-
-  - elastic:
-      name: elastic
-
-  - master:
-      config_file:
-        db:
-          host: localhost
-          port: 5432
-          password: postgres
-          user: postgres
-          name: determined
-        checkpoint_storage:
-          type: s3
-          bucket: hanyu-det-test
-          prefix: /devcluster/checkpoint-download
-        log:
-          level: debug
-        root: tools/build
